Refactor settings save helpers into separate traits
